Move the 'getWorker' function from 'worker-scope.js'
