+++ lib/index.js
@@ -1,4 +1,14 @@
-import getWorker from './worker-scope'
+import workerScope from './worker-scope'
+
+function getWorker (pathWithFilename) {
+  const slash = pathWithFilename.lastIndexOf('/')
+  const path = slash >= 0 ? pathWithFilename.slice(0, slash + 1) : ''
+  const workerInners = `(${workerScope.toString()})('${pathWithFilename}', '${path}')`
+
+  return new Worker(
+    URL.createObjectURL(new Blob([workerInners]))
+  )
+}
 
 export default function wasmWorkerAPI (functions, url) {
   if (!Array.isArray(functions))
