+++ lib/worker-scope.js
@@ -1,14 +1,4 @@
-export default function getWorker (pathWithFilename) {
-  const slash = pathWithFilename.lastIndexOf('/')
-  const path = slash >= 0 ? pathWithFilename.slice(0, slash + 1) : ''
-  const workerInners = `(${workerScope.toString()})('${pathWithFilename}', '${path}')`
-
-  return new Worker(
-      URL.createObjectURL(new Blob([workerInners]))
-    )
-}
-
-function workerScope (pathWithFilename, path) {
+export default function workerScope (pathWithFilename, path) {
   const functions = new Map()
   const heapMap = new Map([
     [  'HEAP8',         Int8Array],
